+++ base/array.jl
@@ -2451,9 +2451,9 @@ function filter!(f, a::AbstractVector)
 end
 
 """
-     mask!(a::AbstractVector, m::AbstractVector{Bool})
+     keepat!(a::AbstractVector, m::AbstractVector{Bool})
 
-The inplace version of logical indexing `a = a[m]`. That is, `mask!(a, m)` on
+The inplace version of logical indexing `a = a[m]`. That is, `keepat!(a, m)` on
 vectors of equal length `a` and `m` will remove all elements from `a` for which
 `m` at the corresponding index is `false`.
 
@@ -2461,7 +2461,7 @@ vectors of equal length `a` and `m` will remove all elements from `a` for which
 ```jldoctest
 julia> a = [:a, :b, :c];
 
-julia> mask!(a, [true, false, true])
+julia> keepat!(a, [true, false, true])
 2-element Vector{Symbol}:
  :a
  :c
@@ -2472,7 +2472,7 @@ julia> a
  :c
 ```
 """
-function mask!(a::AbstractVector, m::AbstractVector{Bool})
+function keepat!(a::AbstractVector, m::AbstractVector{Bool})
     j = firstindex(a)
     for i in eachindex(a, m)
         @inbounds begin